Moved Checkbox and Radio to separate file
